[AWS-TASK-3] - Lambda functions be part
- divide single file into 2 modules
