+++ productService/tsconfig.json
@@ -8,6 +8,8 @@
     "sourceMap": true,
     "resolveJsonModule": true,
     "target": "es2017",
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "outDir": "lib"
   },
   "include": ["./**/*.ts", "../serverless.ts"],
